+++ salesforce/lib/jsontoapex.py
@@ -16,19 +16,22 @@ class JSONConverter():
         from .salesforce.lib.jsontoapex import JSONConverter
         snippet = JSONConverter().convert2apex(name, jsonstr).snippet
     """
-    def __init__(self):
+    def __init__(self, **args):
         self.classes = []
         self.snippet = None
+
+        if "scope" in args:
+            self.scope = args["scope"]
 
     def upcase_first_letter(self, s):
         return s[0].upper() + s[1:]
 
     def add_parser(self, name):
-        parser = "public static {name} parse(String jsonStr) {{\n" + \
+        parser = "{scope} static {name} parse(String jsonStr) {{\n" + \
             "\treturn ({name}) JSON.deserialize(jsonStr, {name}.class);\n" + \
         "}}\n"
 
-        self.classes.append(parser.format(name=name))
+        self.classes.append(parser.format(scope=self.scope, name=name))
 
     def add_testmethod(self, jsonStr):
         testmethod = "static testMethod void testParse() {{\n" + \
@@ -42,7 +45,9 @@ class JSONConverter():
     def convert2apex(self, name, dict_obj, level=0):
         if isinstance(dict_obj, list):
             if len(dict_obj) == 0:
-                self.classes.append("public class %s {\n\n}\n" % name)
+                self.classes.append("%s class %s {\n\n}\n" % (
+                    self.scope, name
+                ))
                 self.snippet = "\n".join(self.classes)[:-1]
                 return self
             else:
@@ -59,29 +64,40 @@ class JSONConverter():
                         data_type = "date"
 
                 if data_type == "dict":
-                    statements.append("\tpublic {type} {name};".format(
-                        type=self.upcase_first_letter(key), name=key
+                    statements.append("\t{scope} {type} {name};".format(
+                        scope=self.scope, 
+                        type=self.upcase_first_letter(key), 
+                        name=key
                     ))
 
                     self.convert2apex(self.upcase_first_letter(key), dict_obj[key], 1)
 
                 elif data_type == "list":
-                    statements.append("\tpublic List<{type}> {name};".format(
-                        type=self.upcase_first_letter(key), name=key
+                    statements.append("\t{scope} List<{type}> {name};".format(
+                        scope=self.scope, 
+                        type=self.upcase_first_letter(key), 
+                        name=key
                     ))
 
                     if not dict_obj[key]:
-                        self.classes.append("public class %s {\n\n}\n" % self.upcase_first_letter(key))
+                        self.classes.append("{scope} class {name} {\n\n}\n".format(
+                            scope=self.scope, 
+                            name=self.upcase_first_letter(key)
+                        ))
                     else:
                         self.convert2apex(self.upcase_first_letter(key), dict_obj[key][0], 1)
 
                 else:
-                    statements.append("\tpublic {type} {name};".format(
-                        type=mappings[data_type], name=key
+                    statements.append("\t{scope} {type} {name};".format(
+                        scope=self.scope, 
+                        type=mappings[data_type], 
+                        name=key
                     ))
 
-            _class = "public class {name} {{\n{class_variable}\n}}\n".format(
-                name=name, class_variable="\n".join(statements)
+            _class = "{scope} class {name} {{\n{class_variable}\n}}\n".format(
+                scope=self.scope, 
+                name=name, 
+                class_variable="\n".join(statements)
             )
 
             if _class not in self.classes: