+++ HISTORY.rst
@@ -4,6 +4,16 @@ Release History
 
 ---------------
 
+
+Release 2.8.8 (2015-04-22)
+++++++++++++++++++
+* Fix bug: If user don't have `Author Apex` privilege, plugin will give wrong information
+* Fix bug: Show alert message if no available package.xml to combine
+* Enhancement: Set ``word_wrap`` of ``Test Run`` result to false
+* Enhancement: Simplify retrieve status check for API version 31 and later, check more detail at `here <https://www.salesforce.com/us/developer/docs/api_meta/Content/meta_retrieve.htm>`_
+* Update documents
+
+
 Release 2.8.7 (2015-04-22)
 ++++++++++++++++++
 * Fix plugin loading NoneType issue